Stabilize JIT faults and procfs task memory access

+++ emu/tlb.h
@@ -15,7 +15,7 @@ struct tlb_entry {
 struct tlb {
     struct mmu *mmu;
     page_t dirty_page;
-    unsigned mem_changes;
+    uint64_t mem_changes;
     // this is basically one of the return values of tlb_handle_miss, tlb_{read,write}, and __tlb_{read,write}_cross_page
     // yes, this sucks
     addr_t segfault_addr;
